Add files via upload
Here are provided the total energy demand on an hourly basis for the past 9 years from march 2008 to December 2018 in the training set we need to forecast the total energy demand on an hourly for the next three years from 2019 to 2021 in the test set. 
Objective- Task to build a machine learning approach to forecast the total energy demand on an hourly basis for next 3 years based on the past trends.